Put in a largish number of signed/unsigned fixes
CVS patchset: 5264
CVS date: 2005/03/04 15:59:08

Mercurial revision: b8cc5248e4884112f27d3a7267bf2528e9fb349e

+++ fencing/test/apitest.c
@@ -37,7 +37,7 @@ stonith_ops_cb(stonith_ops_t * op, void * private_data)
 {
 	printf("optype=%d, node_name=%s, result=%d, node_list=%s\n",op->optype,
 		op->node_name, op->op_result, (char *)op->node_list);
-	if (atoi(private_data) != op->op_result) {
+	if (atoi(private_data) != (int)op->op_result) {
 		g_rc = -1;
 	}
 }
